UC: big rework
drop volley for Ion by @koush
temporary include a custom version, until the repositories got updated.

rework layout, use a cards like optic.
includes a custom app store, which menu is hidden.
it will be unhidden once the server API side is fully ready.

when pressing on an update, it opens the update in detail.
the detailview has a symbol on the top left, indicating its channel.
example: N for Nightly, M for Milestone, RC for Release Client, ...

right of it is the date of the update, below of it the update name

theres also a button, which knows 3 states.
*) downloading, while the update is downloading, give the option to
	cancel it
*) downloaded, allows to automatically install the update
*) not downloaded, allows to download the update

below of it is a (hidden) progressbar, which in future updates shows
the current download progress.

and below that progress bar is a listview, showing the changelog.

when long pressing on an update, it opens the changelog in a dialog.

theres still a lot of work left, but this change brings us a lot closer.

Change-Id: I80e9aadb80e02f98c57aac96b2c120876687dc28